add: events view responsive
