+++ modules/Totem/totem-app/.vite/deps_temp_8ea98a2c/package.json
@@ -1,0 +1,3 @@
+{
+  "type": "module"
+}
