+++ modules/Totem/totem-app/.vite/deps_temp_bf205859/package.json
@@ -1,0 +1,3 @@
+{
+  "type": "module"
+}
