+++ modules/Totem/totem-app/src/config/index.js
@@ -1,5 +1,5 @@
 export const useConfig = () => {
     return Object.freeze({
-        baseURL: 'http://localhost:8080/'
+        baseURL: 'http://192.168.0.22:8080/'
     })
 }
