+++ modules/Totem/totem-app/src/views/EventsView.vue
@@ -2,10 +2,18 @@
 import Header from '@/components/Header.vue'
 import EventCard from '@/components/EventCard.vue'
 import { useEventsStore } from '@/stores/events.js'
+import { useConfig } from '@/config'
 import { computed } from "vue";
+import router from "@/router/index.js";
 
 const eventsStore = useEventsStore()
 const eventsByCategoriesKeys = computed(() => Object.keys(eventsStore.eventsByCategories))
+const config = useConfig()
+
+function logout() {
+  eventsStore.$reset()
+  router.push({name: 'login'})
+}
 
 eventsStore.loadEvents()
 
@@ -22,9 +30,14 @@ eventsStore.loadEvents()
     </div>
 
     <div class="events-wrapper">
-      <div class="load-button-wrapper" v-if="!eventsStore.notHasPageToLoad">
-        <button class="btn btn-primary rounded-5" @click="eventsStore.loadEvents">
+      <div class="actions-buttons-wrapper">
+        <button class="btn btn-primary rounded-5" @click="eventsStore.loadEvents"
+          :class="!eventsStore.notHasPageToLoad ? '' : 'disabled pe-none'">
           <i class="bi bi-plus-circle fs-4"></i>
+        </button>
+
+        <button class="btn btn-dark rounded-5" @click="logout">
+          <i class="bi bi-escape fs-5"></i>
         </button>
       </div>
 
@@ -34,9 +47,9 @@ eventsStore.loadEvents()
             <span class="text-capitalize fs-4">{{category}}</span>
           </div>
 
-          <div class="row row-cols-3 g-3">
+          <div class="row row-cols-1 row-cols-lg-3 g-3">
             <template v-for="(event, index) in eventsStore.eventsByCategories[category]" :key="index">
-              <EventCard :imgSrc="`http://localhost:8080/uploads/events/${event.background}`" :title="event.name"/>
+              <EventCard :imgSrc="`${config.baseURL}uploads/events/${event.background}`" :title="event.name"/>
             </template>
           </div>
         </template>
@@ -64,10 +77,13 @@ eventsStore.loadEvents()
       margin: 1rem 0;
     }
 
-    .load-button-wrapper {
+    .actions-buttons-wrapper {
       position: absolute;
-      bottom: 2rem;
-      right: 5rem;
+      bottom: 10px;
+      right: 25px;
+      display: flex;
+      flex-direction: column;
+      gap: 20px;
     }
   }
 }
